Fixed an issue tracking multiple mouse cursor states
Fixed a problem tracking the previous mouse cursor when doing multiple actions at the same time.
For example, dragging inside a polygon area would previously get the mouse stuck in the mousehand used to indicate the polygon mouseover when dragging ended. Even when the mouse left the polygon, the normal cursor would not be restored because only 1 previous mouse pointer was being tracked.